Refactor Call Kent TTS to full stack component pattern
