+++ app/routes/calls_.record/new.tsx
@@ -2,7 +2,7 @@ import * as React from 'react'
 import { Link, useSearchParams } from 'react-router'
 import { Button, LinkButton } from '#app/components/button.tsx'
 import { CallRecorder } from '#app/components/calls/recorder.tsx'
-import { CallKentTextToSpeech } from '#app/components/calls/text-to-speech.tsx'
+import { CallKentTextToSpeech } from '#app/routes/resources/calls/text-to-speech.tsx'
 import { Grid } from '#app/components/grid.tsx'
 import { Grimmacing } from '#app/components/kifs.tsx'
 import { H4, Paragraph } from '#app/components/typography.tsx'
